Introduce ConsoleState + DeimosConsoleApp wrapper with processor thread
Renames the eframe-facing struct to ConsoleState (state only) and adds a
thin DeimosConsoleApp wrapper that owns Arc<Mutex<ConsoleState>> plus a
dedicated buffer-processor thread. The processor thread runs drain +
health-log + telemetry-tick at 125 Hz (8 ms), decoupling buffer
maintenance from eframe::App::update so freshness holds across repaint
pauses (minimized/occluded window, etc.). render_ui is extracted as an
inherent method on ConsoleState and called by the wrapper's update impl
under the shared mutex. Shutdown is Drop-based: the atomic flips and the
thread joins within one cadence tick. Tasks 1.2 and 1.3 (deleting the
wall-clock-gap arm + reworking tests) intentionally land in later passes.
