--- software/deimos-console/Cargo.toml
@@ -1,28 +1,0 @@
-[package]
-name = "deimos-console"
-version.workspace = true
-edition = "2021"
-authors.workspace = true
-license.workspace = true
-homepage.workspace = true
-description = "Operator console for real-time display of Deimos controller data"
-publish = false
-
-[[bin]]
-name = "deimos-console"
-path = "src/main.rs"
-
-[dependencies]
-eframe = "0.29"
-egui_plot = "0.29"
-serde = { version = "1", features = ["derive"] }
-toml = "0.8"
-clap = { version = "4", features = ["derive"] }
-anyhow = "1"
-crossbeam-channel = "0.5"
-socket2 = { version = "0.5", features = ["all"] }
-chrono = { version = "^0.4.42", features = ["std", "clock"], default-features = false }
-deimos = { path = "../deimos" }
-
-[dev-dependencies]
-tempfile = "3"
