+++ software/deimos_console/README.md
@@ -8,7 +8,7 @@ moment, and displays a connection-health indicator that turns stale when the str
 ## Running
 
 ```sh
-cargo run -p deimos-console -- --config software/deimos-console/examples/console.toml
+cargo run -p deimos_console -- --config software/deimos_console/examples/console.toml
 ```
 
 Run from the repo root. The console connects automatically when the controller enters Operating
@@ -24,7 +24,7 @@ The config file is TOML. All fields except `multicast_group` and `port` are opti
 | `port` | integer | — | UDP port (must match the dispatcher's configured port) |
 | `interface` | IPv4 address | OS default | Local interface address to bind for multicast reception |
 | `window_seconds` | float | `30.0` | Seconds of history kept in each scrolling trace |
-| `staleness_threshold_secs` | float | `2.0` | Seconds without a received row before the connection-health indicator turns stale; also the threshold for the UI-pause and per-row data-staleness signals that drive stall detection |
+| `staleness_threshold_secs` | float | `2.0` | Seconds without a received row before the connection-health indicator turns stale; also the per-row data-staleness threshold (viewer receipt minus controller wall clock) that drives stall detection |
 | `tail_keep_secs` | float | `0.5` | Seconds of pre-stall context retained in each per-channel ring buffer when a stall fires (the rest is evicted so the live window is restored within one repaint) |
 | `recovery_settle_secs` | float | `2.0` | Seconds the connection-health indicator stays in `Recovering` after a stall clears, so an operator notices the discontinuity even on a brief glance |
 | `forensic_log_path` | path | none | If set, a per-session CSV is written here; rotates at 64 MiB |
@@ -49,16 +49,25 @@ log analysis can correlate viewer behavior with the controller's tracing output.
 | `wire_drops` | Cumulative count of sequence-number gaps observed on the wire (controller-side or network packet loss) |
 | `recv_drops` | Cumulative count of messages the receiver thread could not enqueue at all — zero in steady state under drop-oldest semantics |
 | `overwritten_frames` | Cumulative count of receiver-thread drop-oldest evictions; grows under sustained backpressure while preserving freshness |
-| `stalls_detected` | Cumulative count of stall events flagged by the UI-thread heuristic; each event corresponds to one pre-stall buffer clear |
+| `stalls_detected` | Cumulative count of stall events fired when a drained row's lag (viewer receipt minus controller wall clock) exceeds `staleness_threshold_secs`; each event corresponds to one pre-stall buffer clear |
 | `stale_rows_evicted` | Cumulative count of pre-stall samples popped from per-channel ring buffers across all stall events (forensic log retains them) |
 | `pending` | Current depth of the pre-schema pending-rows queue |
 | `lag_ms` | Optional per-row rx-lag triple (viewer receipt minus controller cycle-start wall clock); only present when at least one row was processed this tick |
+
+### Threading
+
+Buffer maintenance runs on a dedicated `deimos-console-processor` thread, not on the egui UI
+thread. That thread ticks every 8 ms (125 Hz) and on each tick drains the receiver channel, runs
+stall detection, evicts pre-stall samples, writes the forensic log, and emits the per-second
+telemetry line. Repaint pacing (whether the window manager calls `eframe::App::update` at 60 Hz,
+slower, or not at all when the window is minimized or occluded) does not affect when the buffer
+moves forward. The cadence is a fixed module constant; there is no config knob.
 
 ### Example
 
 ```toml
 # Sample configuration for the Deimos operator console.
-# Run with: cargo run -p deimos-console -- --config software/deimos-console/examples/console.toml
+# Run with: cargo run -p deimos_console -- --config software/deimos_console/examples/console.toml
 
 # Multicast group the reporting dispatcher publishes to.
 multicast_group = "239.255.0.1"
