+++ software/deimos_console/examples/console.toml
@@ -1,5 +1,5 @@
 # Sample configuration for the Deimos operator console.
-# Run with: cargo run -p deimos-console -- --config software/deimos-console/examples/console.toml
+# Run with: cargo run -p deimos_console -- --config software/deimos_console/examples/console.toml
 
 # Multicast group the reporting dispatcher publishes to.
 multicast_group = "239.255.0.1"
